requirements.txt
pyrtlsdr>=0.2.93
pyModeS>=2.21.1
numpy>=1.26.4